requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
icalendar==4.1.0
idna==3.4
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
urllib3==1.26.12
